+++ package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "history": "^1.13.0",
     "html-to-react": "^0.1.1",
+    "localforage": "^1.3.0",
     "react": "^0.14.1",
     "react-dom": "^0.14.1",
     "react-router": "^1.0.0-rc3",
