+++ src/js/account/UserSettings.js
@@ -1,0 +1,13 @@
+import DataStore from '../utilities/DataStore';
+
+class UserSettings {
+
+    constructor(userObj) {
+
+        this.storage = DataStore.createInstance(userObj.username);
+
+    }
+
+}
+
+export default UserSettings;
