+++ src/js/views/BaseView.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import DataStore from '../utilities/DataStore';
+
+class BaseView extends React.Component {
+
+    constructor(props) {
+        super(props);
+    }
+
+}
+
+export default BaseView;
